src/data-processors/requirements.txt
pyyaml==5.3
glob2==0.7
checksumdir==1.1.7